+++ views/layouts/wiki.php
@@ -15,6 +15,7 @@ $this->title = 'Wiki';
                 ['label' => Yii::t('app', 'OJ 信息'), 'url' => ['wiki/index']],
                 ['label' => Yii::t('app', 'Contest'), 'url' => ['wiki/contest']],
                 ['label' => '出题要求', 'url' => ['wiki/problem']],
+                ['label' => Yii::t('app', 'Special Judge'), 'url' => ['wiki/spj']],
                 ['label' => Yii::t('app', 'About'), 'url' => ['wiki/about']]
             ],
             'options' => ['class' => 'nav nav-pills nav-stacked']