+++ CHANGELOG.md
@@ -12,6 +12,7 @@ JNOJ Change Log
 - Enh: [web] 添加 VIP 用户权限，可以将题目设为只有 VIP 用户可见
 - Bug: [web] 修复 VIP 题目查询问题
 - Bug: [web] 编辑私有题目时无法保存
+- Enh: [web、judge] 对 Special Judge 引入 `testlib.h`，该功能会导致与旧有的 Special Judge 写法不兼容，参考OJ 的 Wiki。
 
 0.7.0 2019.2.1
 ------------------------